_core/master/pyproject.toml
[tool.poetry]
name = "apps-core-master"
version = "0.1.0"
description = "Core master service for managing applications and services"
authors = ["MasterGowen <mastergowen@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
setuptools = ">=82.0.0,<83.0.0"
fastapi = ">=0.129.0,<1.0.0"
uvicorn = {version = ">=0.41.0,<1.0.0", extras = ["standard"]}
nicegui = ">=3.7.1,<4.0.0"
python-keycloak = ">=7.1.1,<8.0.0"
sqlalchemy = ">=2.0.46,<3.0.0"
aiodocker = ">=0.26.0,<1.0.0"
docker = ">=7.1.0,<8.0.0"
aiofiles = ">=25.1.0,<26.0.0"
aiohttp = ">=3.13.3,<4.0.0"
pyyaml = ">=6.0.3,<7.0.0"
jinja2 = ">=3.1.6,<4.0.0"
watchfiles = ">=1.1.1,<2.0.0"
watchdog = ">=6.0.0,<7.0.0"
croniter = ">=6.0.0,<7.0.0"
python-multipart = ">=0.0.22,<1.0.0"
bcrypt = ">=5.0.0,<6.0.0"
requests = ">=2.32.5,<3.0.0"
humanize = ">=4.12.0,<5.0.0"

pydantic-settings = ">=2.13.0,<3.0.0"
[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-asyncio = ">=0.23.0,<1.0.0"
httpx = ">=0.28.1,<1.0.0"
pytest-cov = ">=7.0.0,<8.0.0"
coverage-badge = ">=1.1.2,<2.0.0"

aiosqlite = "^0.22.1"
apprise = "^1.9.9"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "pytest:main"
test-cov = "pytest --cov=app:main"